Rename setting firmware app's update_scenes.* to scenes.*
